review: move splash component to layout directory
